ci(agents): split domain smoke tests from agents
