+++ .github/workflows/agents-domain-ci.yml
@@ -1,0 +1,42 @@
+name: agents-domain-ci
+
+permissions:
+  contents: read
+
+on:
+  pull_request:
+    paths:
+      - 'argocd/applications/jangar-agents-domain/**'
+      - 'argocd/applications/torghut/agents-domain/**'
+      - 'packages/scripts/src/jangar/**'
+      - '.github/workflows/agents-domain-ci.yml'
+  push:
+    branches:
+      - main
+    paths:
+      - 'argocd/applications/jangar-agents-domain/**'
+      - 'argocd/applications/torghut/agents-domain/**'
+      - 'packages/scripts/src/jangar/**'
+      - '.github/workflows/agents-domain-ci.yml'
+
+concurrency:
+  group: ${{ github.workflow }}-${{ github.event.pull_request.number || github.ref }}
+  cancel-in-progress: true
+
+jobs:
+  validate:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Check out repository
+        uses: actions/checkout@v5
+
+      - name: Setup Bun
+        uses: oven-sh/setup-bun@v2
+        with:
+          bun-version: 1.3.13
+
+      - name: Install dependencies
+        run: bun install --frozen-lockfile --ignore-scripts --filter @proompteng/scripts
+
+      - name: Run Agents domain smoke tests
+        run: bun test packages/scripts/src/jangar/__tests__/agents-domain-smoke.test.ts
