+++ services/jangar/src/server/dev.ts
@@ -1,13 +1,13 @@
-import { installAgentsEnvCompatibility } from '@proompteng/agents/server/env-compat'
 import { createServer as createViteServer } from 'vite'
 
+import { installJangarEnvCompatibility } from './env-compat'
 import { bootRuntimeProfile } from './runtime-boot'
 import { JANGAR_RUNTIME_PROFILES } from './runtime-profile'
 import { resolveHttpServerListenConfig } from './runtime-entry-config'
 
 type JangarRuntime = Awaited<ReturnType<typeof import('./app').createJangarRuntime>>
 
-installAgentsEnvCompatibility()
+installJangarEnvCompatibility()
 
 const { port, hostname, idleTimeoutSeconds } = resolveHttpServerListenConfig(process.env, { dev: true })
 const runtimeProfile = JANGAR_RUNTIME_PROFILES.viteDevApi
